Moved integrations settings page js into teamSettings.js
