+++ scripts/application/application.js
@@ -41,4 +41,3 @@ require('../shared/checkAll');
 require('../shared/workflowSettings');
 require('../shared/contentHeader');
 require('../shared/actionBar');
-require('../shared/integrations');
